requirements.txt
pyxmpp==1.1.1
yaml==3.09
